separate __main__ from burst.py
